feat: Refactor after rebase
