+++ posthog/admin/admins/__init__.py
@@ -12,7 +12,7 @@ from .person_distinct_id_admin import PersonDistinctIdAdmin
 from .plugin_admin import PluginAdmin
 from .plugin_config_admin import PluginConfigAdmin
 from .project_admin import ProjectAdmin
-from .redis_admin import RedisAdmin
+from .redis_mutation_admin import RedisMutationAdmin
 from .survey_admin import SurveyAdmin
 from .team_admin import TeamAdmin
 from .text_admin import TextAdmin
